StatisticsParams moved to proper folders
