+++ apps/vaporgui/CMakeLists.txt
@@ -47,7 +47,6 @@ set (SRCS
 	ImageEventRouter.cpp
 	ContourEventRouter.cpp
 	PlotParams.cpp
-	StatisticsParams.cpp
 	MiscParams.cpp
 	ContourSubtabs.cpp
 	BarbSubtabs.cpp
