+++ lib/params/CMakeLists.txt
@@ -20,6 +20,7 @@ add_library (params SHARED
 	ImageParams.cpp
 	TFInterpolator.cpp
 	ContourParams.cpp
+	StatisticsParams.cpp
 )
 
 if (${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
